Extract user update to component file
