+++ client/src/components/admin/UserData.js
@@ -3,12 +3,12 @@ import AnalyticsContext from "../../context/analytics/AnalyticsContext"
 
 import { CSSTransition } from "react-transition-group"
 
-import { createUserSnapshot, getAllUsers, submitUserInfoUpdates, sumUsers } from "../../utils/api/analytics"
+import { createUserSnapshot, getAllUsers, sumUsers } from "../../utils/api/analytics"
 import { getLocaleDate } from "../../utils/dates"
 
 import { CircularProgress } from "@material-ui/core"
-import CloseOutlinedIcon from "@material-ui/icons/CloseOutlined"
 import PageNav from "../layout/PageNav"
+import UserUpdate from "./UserUpdate"
 
 const UserData = () => {
   const analyticsContext = useContext(AnalyticsContext)
@@ -69,36 +69,6 @@ const UserData = () => {
     setUserPopup(true)
   }
 
-  const [newUserData, setNewUserData] = useState({})
-
-  useEffect(() => {
-    if (currentUser) setNewUserData({ admin: currentUser.admin })
-    else setUserPopup(false)
-  }, [currentUser])
-
-  const updateUserInfo = e => {
-    const { name, value } = e.target
-    if (value) {
-      setNewUserData(prev => ({ ...prev, [name]: value }))
-    } else {
-      setNewUserData(prev => {
-        let newData = { ...prev }
-        delete newData[name]
-        return newData
-      })
-    }
-  }
-
-  const toggleAdmin = e => {
-    setNewUserData(prev => ({ ...prev, admin: e.target.checked }))
-  }
-
-  const submitUserInfo = e => {
-    e.preventDefault()
-    submitUserInfoUpdates({ updates: newUserData, _id: currentUser._id }, dispatch)
-    setNewUserData({})
-  }
-
   return (
     <div className="analytics-content">
       <div>
@@ -137,42 +107,7 @@ const UserData = () => {
         <PageNav loading={loading} resultsPerPage={resultsPerPage} totalRecords={totalUsers} changePages={changePages} />
       </div>
       <CSSTransition in={userPopup} classNames="modal-content" timeout={150} unmountOnExit>
-        <div className="user-action__container">
-          <div className="user-action">
-            <div className="user-action--user">
-              <h2>{currentUser?.name}</h2>
-              <p>{currentUser?.email}</p>
-              {/* <p>{currentUser?.lists.length}</p> */}
-              <p className="user-activity">
-                Last login: {`${Math.floor((new Date() - new Date(currentUser?.lastLogin)) / 1000 / 60 / 60 / 24)}`} days ago
-              </p>
-              <p className="badge">{currentUser?.admin ? "Admin" : "User"}</p>
-              <button className="close-user-popup" onClick={() => setUserPopup(false)}>
-                <CloseOutlinedIcon />
-              </button>
-            </div>
-            <form className="user-update-form" onSubmit={submitUserInfo}>
-              <h3>Update user info</h3>
-              <label>
-                New Name
-                <input type="text" value={newUserData.name || ""} onChange={updateUserInfo} name="name" />
-              </label>
-              <label>
-                New Email
-                <input type="email" value={newUserData.email || ""} onChange={updateUserInfo} name="email" />
-              </label>
-              <label>
-                New Password
-                <input type="password" value={newUserData.password || ""} onChange={updateUserInfo} name="password" />
-              </label>
-              <label>
-                {newUserData.admin ? "Revoke Admin" : "Grant Admin"}
-                <input type="checkbox" onChange={toggleAdmin} checked={newUserData.admin || false} name="admin" />
-              </label>
-              <button type="submit">Update User</button>
-            </form>
-          </div>
-        </div>
+        <UserUpdate setUserPopup={setUserPopup} currentUser={currentUser} />
       </CSSTransition>
     </div>
   )
